Consolidate .netcoreapp.cs files because System.Reflection project is no longer cross-compiled
