+++ src/libraries/System.Reflection/tests/AssemblyTests.cs
@@ -24,6 +24,8 @@ TypeAttr(typeof(object), name = "TypeAttrSimple")]
 [assembly: CompilationRelaxations(8)]
 [assembly: Debuggable((DebuggableAttribute.DebuggingModes)263)]
 [assembly: CLSCompliant(false)]
+[assembly:TypeForwardedTo(typeof(string))]
+[assembly: TypeForwardedTo(typeof(TypeInForwardedAssembly))]
 
 namespace System.Reflection.Tests
 {
@@ -748,6 +750,46 @@ namespace System.Reflection.Tests
         {
             IEnumerable<CustomAttributeData> customAttributesData = typeof(AssemblyTests).Assembly.GetCustomAttributesData().Where(cad => cad.AttributeType == attrType);
             Assert.True(customAttributesData.Count() > 0, $"Did not find custom attribute of type {attrType}");
+        }
+
+        [Fact]
+        public static void AssemblyGetForwardedTypes()
+        {
+            Assembly a = typeof(AssemblyNetCoreAppTests).Assembly;
+            Type[] forwardedTypes = a.GetForwardedTypes();
+
+            forwardedTypes = forwardedTypes.OrderBy(t => t.FullName).ToArray();
+
+            Type[] expected = { typeof(string), typeof(TypeInForwardedAssembly), typeof(TypeInForwardedAssembly.PublicInner), typeof(TypeInForwardedAssembly.PublicInner.PublicInnerInner) };
+            expected = expected.OrderBy(t => t.FullName).ToArray();
+
+            Assert.Equal<Type>(expected, forwardedTypes);
+        }
+
+        [Fact]
+        public static void AssemblyGetForwardedTypesLoadFailure()
+        {
+            Assembly a = typeof(TypeInForwardedAssembly).Assembly;
+            ReflectionTypeLoadException rle = Assert.Throws<ReflectionTypeLoadException>(() => a.GetForwardedTypes());
+            Assert.Equal(2, rle.Types.Length);
+            Assert.Equal(2, rle.LoaderExceptions.Length);
+
+            bool foundSystemObject = false;
+            bool foundBifException = false;
+            for (int i = 0; i < rle.Types.Length; i++)
+            {
+                Type type = rle.Types[i];
+                Exception exception = rle.LoaderExceptions[i];
+
+                if (type == typeof(object) && exception == null)
+                    foundSystemObject = true;
+
+                if (type == null && exception is BadImageFormatException)
+                    foundBifException = true;
+            }
+
+            Assert.True(foundSystemObject);
+            Assert.True(foundBifException);
         }
 
         private static Assembly LoadSystemCollectionsAssembly()
